Refactors EditorView to use UI panels
Replaces conditional compilation in EditorView with UI panels for
feature management.

This change enables dynamic feature visibility based on availability,
leading to a cleaner codebase and improved maintainability. It centralizes
feature-specific logic within dedicated panel components, reducing
code complexity and improving testability.

Removes the refactoring guide.

+++ Source/TeleprompterRenderStrategy.cpp
@@ -167,7 +167,7 @@ float TeleprompterRenderStrategy::calculateScrollOffset (const std::vector<LineI
                                                           float readLineY) const
 {
     // Find the line containing the current word from the current clip
-    for (int lineIndex = 0; lineIndex < allLines.size(); ++lineIndex)
+    for (size_t lineIndex = 0; lineIndex < allLines.size(); ++lineIndex)
     {
         const auto& line = allLines[lineIndex];
